requirements.txt
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
black==25.1.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
Django==5.2
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
drf-yasg==1.21.10
filelock==3.18.0
identify==2.6.9
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
model-bakery==1.20.4
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
ptyprocess==0.7.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
virtualenv==20.30.0
